# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NSE/BSE AI Trading Signals API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get all current AI trading signals
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [STOCK, OPTIONS, FUTURES, ALL]
            default: ALL
        - name: action
          in: query
          schema:
            type: string
            enum: [BUY, SELL, EXIT, ALL]
            default: ALL
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, TRIGGERED, EXPIRED, ALL]
            default: ACTIVE
      responses:
        "200":
          description: Trading signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSignal"
  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a specific signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingSignal"
        "404":
          description: Not found
  /signals/generate:
    post:
      operationId: generateSignals
      tags: [signals]
      summary: Trigger AI agent to generate new trading signals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSignalsInput"
      responses:
        "200":
          description: Newly generated signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingSignal"
...
  /openai/agent/analyze:
    post:
      operationId: runAgentAnalysis
      tags: [openai]
      summary: Run the AI agent to analyze a symbol and produce a detailed report with signals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunAgentAnalysisInput"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentAnalysisResult"
components:
  schemas:
...
    TradingSignal:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        instrumentType:
          type: string
          enum: [STOCK, OPTIONS, FUTURES]
        action:
          type: string
          enum: [BUY, SELL, EXIT]
        displayText:
          type: string
          description: Human-readable signal e.g. "BUY NIFTY 22000 CE"
        entryPrice:
          type: number
          nullable: true
        targetPrice:
          type: number
          nullable: true
        stopLoss:
          type: number
          nullable: true
        confidence:
          type: number
          description: 0-100
        rationale:
          type: string
        status:
          type: string
          enum: [ACTIVE, TRIGGERED, EXPIRED]
        timeframe:
          type: string
          enum: [INTRADAY, SWING, POSITIONAL]
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required: [id, symbol, instrumentType, action, displayText, confidence, rationale, status, timeframe, createdAt]
    GenerateSignalsInput:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
          description: List of symbols to analyze; empty means use watchlist
        timeframe:
          type: string
          enum: [INTRADAY, SWING, POSITIONAL]
          default: INTRADAY
        provider:
          type: string
          default: fallback
          description: Preferred AI provider / LLM engine to generate signals
      required: []
...
    RunAgentAnalysisInput:
      type: object
      properties:
        symbol:
          type: string
        instrumentType:
          type: string
          enum: [STOCK, OPTIONS, FUTURES]
          default: STOCK
        timeframe:
          type: string
          enum: [INTRADAY, SWING, POSITIONAL]
          default: INTRADAY
        numSignals:
          type: integer
          default: 2
        maxTokens:
          type: integer
          default: 2048
        style:
          type: string
          enum: [conservative, moderate, aggressive]
          default: moderate
        customContext:
          type: string
          default: ""
        confidenceThreshold:
          type: integer
          default: 50
        saveSignals:
          type: boolean
          default: true
        provider:
          type: string
          default: fallback
      required: [symbol]
    AgentAnalysisResult:
      type: object
      properties:
        symbol:
          type: string
        summary:
          type: string
        keyLevels:
          type: object
          properties:
            support:
              type: array
              items:
                type: number
            resistance:
              type: array
              items:
                type: number
          required: [support, resistance]
        signals:
          type: array
          items:
            $ref: "#/components/schemas/TradingSignal"
        riskAssessment:
          type: string
        generatedAt:
          type: string
      required: [symbol, summary, keyLevels, signals, riskAssessment, generatedAt]
...