# lib/api-zod/src/generated/types/generateSignalsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NSE/BSE AI Trading Signals API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateSignalsInputTimeframe } from "./generateSignalsInputTimeframe";

export interface GenerateSignalsInput {
  /** List of symbols to analyze; empty means use watchlist */
  symbols?: string[];
  timeframe?: GenerateSignalsInputTimeframe;
  /** Preferred AI provider / LLM engine to generate signals */
  provider?: string;
}